Set up passport and refactor routes
Routes have been reorganized and now support passport. Only a local
strategy is implemented, but this will help later on when implementing
the services that Dash will support. Passport was the remaining piece to
implement after MongoDB and the user model was set up. The user could be
created and modified, but no persistent session was kept of that user
upon logging in. Passport completes the authentication process.

The following have been tested and are working:

- User login
- User registration
- Default display name created from email address (the portion before
  the '@' symbol)
- Account 2-hour lock upon 5 failed login attempts
- Form error messages using the connect-flash package
- Passport's handling of a persistent session for the user
- Check if user is logged in before visiting pages that require the user
  to be logged in
- Logout

The following will be implemented in the future:

- 'Forgot your password'
- Email notification for account lock
- Email verification
- Display name changes (through a user settings page)