Update connection process
As opposed to configuring your settings in one go, users will instead
connect with the service of their choice and return to the connect page
with a message indicating whether the connection was successful. They
are then able to configure specific settings.

This method bypasses the need for flags to enforce going through the
settings process and instead creates single-step operations. The user
doesn't have to go from (1) Facebook connect to (2) Facebook groups to
(3) Facebook pages and back to (4) the connect page. The user could
leave at any time and lose their place in the process. Flags would only
make the backend more complex.

Instead, only steps 1 and 4 occur in a single operation: adding the
connection. This turns the 1-2-3-4 cycle into various single
step-operations as follows: 1-4, 2-4, 3-4. The user connects or
modifies settings and is returned to the connect page. There is no
room for accidental leave or error - if an error does occur, their
settings remain intact and they return to the connect page with a
proper error.