Complete implementation for new service: YouTube
This closes #9. Since the number of supported services will continue to
increase, the stylesheet has been modified to mark all cards on the
dashboard by the 'card-dashboard' class. 'card-[service name here]' will
still be used to target cards belonging to a specific service. However,
all styles related to the cards on the dashboard will target
'card-dashboard' to avoid redundancy as well as prevent conflicts with
other cards in the setup pages.