Split updates into batches of posts
This resolves #16 in revising the schema to split posts into their own
updates such that the user can follow the update timeline. Previously,
the user would possible see older posts at the top of the list, due to a
user-triggered reset or update on the connections page. Now, we sort by
update times. The posts themselves are in chronological order, but the
updates go in order of the user's actions. So, if the user decides to
manually update YouTube after resetting the update time, the posts
within that update will be in chronological order and the action of only
updating YouTube will come up on the top, as it was the most recent.