Clean up dashboard pagination and fix new bugs
The number of pages should be used instead of the items per page for
calculating the start and end pages and where to insert breaks (i.e.
'...'). The page number now only appears if we're not on the first page.
The pagination has been moved to its own function to avoid redundancy.