Refactor user model and add simple mongoose test
The user model has been refactored to allow for independent inclusion
into Mocha tests. Using 'require.main.require' is an issue for when
mocha is executed as opposed to the app, since Mocha cannot reference
all of the dependencies. The model dependencies for the user model have
been moved from app.js into the user model file. The messages and config
are now passed as parameters to services. This allows for Mocha to
require the user model and complete simple tests.